04-20-2026/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
lxml==6.1.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.1
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
six==1.17.0
svgpath2mpl==1.0.0
typing_extensions==4.15.0
xlsxwriter==3.2.9
